Desenvolvimento

+++ FPTV/Models/StatisticsModels/MatchPlayerStatsVal.cs
@@ -41,7 +41,7 @@ namespace FPTV.Models.StatisticsModels
 
         [Required]
         [Display(Name = "ADR")]
-        public float? ADR { get; set; }
+        public double? ADR { get; set; }
 
         [Required]
         [Display(Name = "Kast")] 
@@ -49,11 +49,11 @@ namespace FPTV.Models.StatisticsModels
 
         [Required]
         [Display(Name = "HeadShots")] 
-        public float? HeadShots { get; set; }
+        public double? HeadShots { get; set; }
 
         [Required]
         [Display(Name = "KD_Diff")] 
-        public float? KD_Diff { get; set; }
+        public double? KD_Diff { get; set; }
 
         [Required]
         [Display(Name = "Name of a player")]
